requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==3.2.7
django-appconf==1.0.5
django-filter==21.1
django-flat-responsive==2.0
django-heroku==0.3.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-simple-history==3.0.0
gunicorn==20.1.0 
mysqlclient==2.0.3
phonenumberslite==8.12.35
Pillow==8.4.0
psycopg2==2.9.2
python-decouple==3.5
pytz==2021.1
reportlab==3.6.5
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
whitenoise==5.3.0
